cheri riscv for 5.15.48
This is a CHERI RISCV support patch for the Linux kernel.
It enables execution of CHERI purecap userspace programs,
as well as maintains support for normal programs.

It contains minimal changes required to support CHERI in user space:

1. Cheri enabled context switches
2. Cheri enabled exception handling
3. Cheri enabled signals
4. And various changes around kernel to support Cheri userspace capabilities.
5. etc...

This is an experimental code and work-in-progress.
The code is written with kernel codying style guideline in mind.
But as it work-in-progress, it may contain some odds things :)

Written by:
Dmitry Kasatkin <dmitry.kasatkin@huawei.com>
Kui Wang <wang.kui1@huawei.com>

+++ arch/riscv/include/asm/cmpxchg.h
@@ -316,7 +316,7 @@
 			"	fence rw, rw\n"				\
 			"1:\n"						\
 			: "=&r" (__ret), "=&r" (__rc), "+A" (*__ptr)	\
-			: "rJ" ((long)__old), "rJ" (__new)		\
+			: "rJ" ((long)__old), "rJ" ((long)__new)	\
 			: "memory");					\
 		break;							\
 	case 8:								\
